+++ a5/a5/Player.h
@@ -8,8 +8,8 @@
 class Floor;
 class Player : public Creature {
 protected:
-	int atkMod;			//Potion caused buff/debuff to attack for current floor
-	int defMod;			//Potion caused buff/debuff to defense for current floor
+	double atkMod;			//Potion caused buff/debuff to attack for current floor
+	double defMod;			//Potion caused buff/debuff to defense for current floor
 	int chamber;		//The current chamber the Player is in
 	int gold;			//Player's current gold
 	bool knownPots[6];	//Potions discovered by the player character
@@ -19,7 +19,7 @@ public:
 	Player(int, int, int, Floor*);
 	virtual ~Player() = 0;
 	int getChamber();
-	virtual int getDefense();
+	virtual double getDefense();
 	virtual void attack(Creature *);
 	virtual bool movement(std::string);
 	virtual bool consumePotion(std::string);
